Move user pool data into separate component
